Move TestScript to ilsgateway
