+++ custom/ilsgateway/tests/test_report_reminders.py
@@ -14,9 +14,9 @@ from custom.ilsgateway.tanzania.reminders.reports import get_district_people, co
     construct_delivery_summary, construct_soh_summary
 from custom.ilsgateway.tests.handlers.utils import prepare_domain
 from custom.ilsgateway.tests.handlers.utils import TEST_DOMAIN, create_products
-from custom.ilsgateway.utils import make_loc
-from custom.logistics.tests.test_script import TestScript
+from custom.ilsgateway.tests.test_script import TestScript
 from custom.ilsgateway.tests.utils import bootstrap_user
+from custom.ilsgateway.utils import make_loc
 import six
